+++ utbot-junit-contest/src/main/kotlin/org/utbot/contest/ContestEstimator.kt
@@ -148,8 +148,8 @@ enum class Tool {
                     fuzzingRatio,
                     project.sootClasspathString,
                     runFromEstimator = true,
-                    methodNameFilter,
-                    expectedExceptions
+                    expectedExceptions,
+                    methodNameFilter
                 )
             } catch (e: CancellationException) {
                 logger.info { "[$classFqn] finished with CancellationException" }
@@ -410,7 +410,7 @@ fun runEstimator(
         val project = ProjectToEstimate(
             name,
             classesFQN,
-            projectToExpectedExceptions[name]!!,
+            projectToExpectedExceptions.getValue(name),
             File(classpathDir, name).listFiles()!!.filter { it.toString().endsWith("jar") },
             testCandidatesDir,
             unzippedJars
@@ -533,57 +533,6 @@ private fun classNamesByJar(jar: File): List<String> {
             className.substringBeforeLast(".class")
         }
         .toList()
-}
-
-class ExpectedExceptionsForMethod(
-    val exceptionNames: List<String>
-)
-
-class ExpectedExceptionsForClass {
-
-    private val byMethodName: MutableMap<String, ExpectedExceptionsForMethod> =
-        mutableMapOf()
-
-    fun getForMethod(methodName: String): ExpectedExceptionsForMethod =
-        byMethodName[methodName] ?: ExpectedExceptionsForMethod(listOf())
-
-    fun setForMethod(methodName: String, exceptionNames: List<String>) {
-        byMethodName[methodName] = ExpectedExceptionsForMethod(exceptionNames)
-    }
-}
-
-class ExpectedExceptionsForProject {
-
-    private val byClassName: MutableMap<String, ExpectedExceptionsForClass> =
-        mutableMapOf()
-
-    fun getForClass(className: String): ExpectedExceptionsForClass =
-        byClassName[className] ?: ExpectedExceptionsForClass()
-
-    fun setForClass(className: String, methodName: String, exceptionNames: List<String>) {
-        val forClass = byClassName.getOrPut(className) { ExpectedExceptionsForClass() }
-        forClass.setForMethod(methodName, exceptionNames)
-    }
-}
-
-fun parseExceptionsFile(exceptionsDescriptionFile: File): ExpectedExceptionsForProject {
-    if (!exceptionsDescriptionFile.exists()) {
-        return ExpectedExceptionsForProject()
-    }
-
-    val forProject = ExpectedExceptionsForProject()
-    for (methodDescription in exceptionsDescriptionFile.readLines()) {
-        val methodFqn = methodDescription.substringBefore(':').trim()
-        val classFqn = methodFqn.substringBeforeLast('.')
-        val methodName = methodFqn.substringAfterLast('.')
-        val exceptionFqns = methodDescription.substringAfter(':')
-            .split(' ')
-            .map { it.trim() }
-            .filter { it.isNotBlank() }
-        forProject.setForClass(classFqn, methodName, exceptionFqns)
-    }
-
-    return forProject
 }
 
 class ProjectToEstimate(
